fraud_detection_backend/requirements.txt
Flask
celery
redis
psycopg2-binary
SQLAlchemy
pandas
scikit-learn
python-dateutil
rethinkdb
python-dotenv